requirements.txt
azure-cosmos
python-dotenv
fastapi
uvicorn
aiohttp